LayoutTests/fast/hidpi/image-srcset-data-src.html

<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script>
    addEventListener("load", function() {
        shouldBeTrue('document.getElementById("foo").clientWidth==150');
    }, false);
</script>
</head>

<body id="body">
    <div>This test passes if the image below looks like a 2x2 grid (yellow, gray, light green, dark green).</div>
    <img id="foo" src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAJYAAAA8CAIAAAAL5NQ9AAAACXBIWXMAAAsTAAALEwEAmpwYAAAAB3RJTUUH2woaBQc4oLEFpAAAABl0RVh0Q29tbWVudABDcmVhdGVkIHdpdGggR0lNUFeBDhcAAACMSURBVHja7dNBEYAgFEVRPhHMYgAzUIsmVnFvB/fsoQb+ObfBmzMvxneW1D1vzz2wFiEUQiFEKIRCKIQIhVAIhRChEAqhECIUQiEUQoRCKIRCiFAIhVAIEep3xTWTLzzu5oVCKIRCiFAIhVAIEQqhEAohQiEUQiFEKIRCKIQIhVAIhRChEAqhECLUZi3VEwcBMGr1NgAAAABJRU5ErkJggg==">
</body>
</html>
