third_party/WebKit/LayoutTests/fast/hidpi/image-srcset-data-src-expected.txt

PASS document.getElementById("foo").clientWidth==150 is true
PASS successfullyParsed is true

TEST COMPLETE
This test passes if the image below looks like a 2x2 grid (yellow, gray, light green, dark green).
